requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib.pyplot
joblib
scikit-learn